10fold validation for ML method
it's a script write for one of my project. it will test the prediction true rate for 10 times validation.
but some of arguments are writen in the code. it should be change to suit next dataset. 
if i have more time, i will fix this problem, and make it more convenient and applicability.